Can't limit point QoI assembly to 1 element
Even with \kappa \le 1, it's easy for it to spill over to another
element. For now, we're back to integrating over every element
(expensive for higher order quadrature rules), but we ought to
be able to precompute a patch over which to do the QoI assembly.
Put the option in for that, just need to actually implement it.