Change sign convention in solid mechanics Physics
Originally I'd written these thinking something liek M\ddot{u} = F(u).
But, really, what we want is a residual
M\ddot{u} + C\dot{u} + F(u) = 0 (or the minus of this, which is
what we have before).
So this updates the sign convention for all the solid mechanics related
classes to conform to this.