Mostly hardcode to 2D case, first order time for displacement
Out of desperation, we're clearing out all the crap related to 3D,
and to trying to preserve second order in time displacment variables.
Instead, we're hardcoding to 2D, and making displacement variables
first order in time so that we can eliminate that juggling for now.
Only remaining part is the Jacobian of the fluid source term. The
hardcode-to-2D changes are mixed in with finite differencing the
Jacobian.