FIX MAJOR ERROR: reinit takes *REFERENCE* coordinates
The coordinates of the displaced quadrature points are in physical
element coordinates, but FEMContext::elem_fe_reinit (which internally
calls FEAbstract::reinit, the main function) is expecting coordinates
in the REFERENCE element. So why didn't this give us screaming errors
before? Because all our test cases had the solid located somewhere
within the [-1.1]x[-1,1] box in the physical domain, but that also
happens to be the bounds for reference element coordinates.